guard/requirements.txt
PyYAML~=6.0.1
typing~=3.7.4.3
logging~=0.4.9.6
pathlib~=1.0.1
uvicorn~=0.38.0
pydantic~=2.12.4
sqlalchemy~=2.0.44
alembic~=1.17.2
python-dotenv~=1.2.1
fastapi~=0.121.3
asyncpg~=0.30.0
psycopg2-binary~=2.9.9
pytest~=8.0.0
pytest-asyncio~=0.23.0
httpx~=0.27.0
pip~=25.0.1
attrs~=23.2.0
redis~=7.1.0
PyJWT~=2.9.0
passlib~=1.7.4
bcrypt~=4.0.1

